Fix backward of binary_cross_entropy_with_logits
The previous PR in this stack uncovered an error in the forward over
backward for this function.

In this PR, we fix this error and we also fix the gradgrad
implementation (and make it more stable and faster using `logsigmoid`).
We also move the double backward for this function to `FunctoinsManual`
as there's no reason for it to be in `native_functions`

ghstack-source-id: 9357ec9463f0a99a5b8d89e532ad2ef4ef78bd2f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79381